+++ src/Parse.js
@@ -18,6 +18,7 @@ import RESTController from './RESTController';
 import MoralisWeb3 from './MoralisWeb3';
 const { checkForSdkUpdates } = require('./utils');
 import { ethers } from 'ethers';
+import { filterConsole } from './filterConsole';
 
 /**
  * Contains all Moralis API classes and functions.
@@ -309,7 +310,11 @@ Moralis.Elrond = require('./MoralisErd').default;
 Moralis.Erd = Moralis.Elrond;
 Moralis.Dot = require('./MoralisDot').default;
 Moralis.UI = require('./MoralisUI').default;
-Moralis.Chains = require('./Chians').default;
+Moralis.Chains = require('./Chains').default;
+Moralis.AbstractWeb3Connector = require('./Web3Connector/AbstractWeb3Connector').default;
+
+filterConsole();
+
 Moralis._request = function (...args) {
   return CoreManager.getRESTController().request.apply(null, args);
 };